fix: widen DMA userParam to uintptr_t; test IRQ handler round-trip
dmaChannelDescriptor_t::userParam was uint32_t, truncating the extDevice_t*
that bus_spi_ll.c/bus_spi_stdperiph.c's spiInternalStartDMA() stores into
it via (uint32_t)dev, on any host wider than 32 bits (UNIT_TEST).
bus_spi.c's spiRxIrqHandler/spiTxIrqHandler read the same field back and
cast it to a pointer, so a truncated value there reconstructs a wrong
pointer, not just a compiler warning. CodeRabbit's review pointed at the
read-side casts (bus_spi.c); the actual truncating write happens in
bus_spi_ll.c/bus_spi_stdperiph.c, confirmed by tracing every writer of
->userParam before fixing.

Widens the field and dmaSetHandler()'s userParam parameter (dma.h, dma.c,
dma_stm32f4xx.c, dma_stm32f7xx.c, dma_stm32h7xx.c) to uintptr_t, and the
two write-side casts (bus_spi_ll.c, bus_spi_stdperiph.c) from uint32_t to
uintptr_t. No-op on the real 32-bit target. The two read-side casts in
bus_spi.c drop their now-redundant intermediate uintptr_t cast.

Adds two tests that capture the real spiRxIrqHandler/spiTxIrqHandler via
the mock dmaSetHandler(), write a device pointer through userParam, and
invoke the captured handler directly -- exercising the round-trip
CodeRabbit's review asked for instead of asserting on the field's type.

Verified: 42/42 test binaries pass (249 cases, up from 243); HELIOSPRING,
FOXEERF722V4, STELLARH7DEV, and SITL all build clean with no new warnings.

+++ src/main/drivers/bus_spi.c
@@ -255,8 +255,7 @@ FAST_CODE static void spiIrqHandler(const extDevice_t *dev)
 // Negates CS, stops DMA, invalidates cache on F7, then advances the segment list.
 FAST_CODE static void spiRxIrqHandler(dmaChannelDescriptor_t *descriptor)
 {
-    // uintptr_t avoids a -Wint-to-pointer-cast error under -Werror on 64-bit hosts (UNIT_TEST).
-    const extDevice_t *dev = (const extDevice_t *)(uintptr_t)descriptor->userParam;
+    const extDevice_t *dev = (const extDevice_t *)descriptor->userParam;
 
     if (!dev) {
         return;
@@ -286,8 +285,7 @@ FAST_CODE static void spiRxIrqHandler(dmaChannelDescriptor_t *descriptor)
 // DMA Tx TC IRQ handler — used for Tx-only DMA buses (no Rx channel).
 FAST_CODE static void spiTxIrqHandler(dmaChannelDescriptor_t *descriptor)
 {
-    // uintptr_t avoids a -Wint-to-pointer-cast error under -Werror on 64-bit hosts (UNIT_TEST).
-    const extDevice_t *dev = (const extDevice_t *)(uintptr_t)descriptor->userParam;
+    const extDevice_t *dev = (const extDevice_t *)descriptor->userParam;
 
     if (!dev) {
         return;
